refactor file structure to give each component a file
