+++ src/components/NewListButton.js
@@ -1,0 +1,13 @@
+import React from 'react'
+
+const NewListButton = ( {handleNewListButton} ) => {
+  return (
+    <li className='newListButton'>
+      <span className='newListButtonSpan' onClick={handleNewListButton}>{'[＋]'}</span>
+    </li>
+  )
+}
+
+export default NewListButton
+
+
