+++ src/components/SettingsButton.js
@@ -1,0 +1,13 @@
+import React from 'react'
+
+const SettingsButton = () => {
+    return (
+      <div className="settingsButtonWrapper">
+        <a href="https://nico.gl" className="float">
+            Made by Nico ✨
+        </a>
+      </div>
+    )
+  }
+
+  export default SettingsButton
